+++ CHANGELOG.md
@@ -5,9 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-<!-- _No unreleased changes_
+<!-- _No unreleased changes_ -->
 
-## [1.2.0] - 2020-08-31 -->
+## [1.2.0] - 2020-08-31
 ### Added 
 - Added run 90 second ad button
 - Added create stream marker button
@@ -78,7 +78,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2020-06-25
 ### Initial Release
 
-<!-- [1.2.0]: https://github.com/VariXx/varibot-twitch-js/tree/v1.2.0 -->
+[1.2.0]: https://github.com/VariXx/varibot-twitch-js/tree/v1.2.0
 [1.1.1]: https://github.com/VariXx/varibot-twitch-js/tree/v1.1.1
 [1.1.0]: https://github.com/VariXx/varibot-twitch-js/tree/v1.1.0
 [1.0.3]: https://github.com/VariXx/varibot-twitch-js/tree/v1.0.3